+++ src/main/java/nextstep/reservation/service/ReservationService.java
@@ -6,9 +6,9 @@ import nextstep.reservation.dto.ReservationRequest;
 import nextstep.reservation.repository.ReservationDao;
 import nextstep.schedule.Schedule;
 import nextstep.schedule.repository.ScheduleDao;
-import nextstep.support.DuplicateEntityException;
-import nextstep.support.NotExistEntityException;
-import nextstep.support.UnauthorizedAccessException;
+import nextstep.exception.DuplicateEntityException;
+import nextstep.exception.NotExistEntityException;
+import nextstep.exception.UnauthorizedAccessException;
 import nextstep.theme.Theme;
 import nextstep.theme.repository.ThemeDao;
 import org.springframework.stereotype.Service;
